requirements.txt
duckdb
pyarrow
pathlib
tensorboard
Pillow
